rename view to view_function
